Splits history to one store per file
