+++ frontend/app/src/store/session/index.ts
@@ -24,8 +24,9 @@ import { useBalancesStore } from '@/store/balances';
 import { useEthNamesStore } from '@/store/balances/ethereum-names';
 import { Section, Status } from '@/store/const';
 import { useDefiStore } from '@/store/defi';
-import { useHistory, useTransactions } from '@/store/history';
+import { useHistory } from '@/store/history';
 import { useTxQueryStatus } from '@/store/history/query-status';
+import { useTransactions } from '@/store/history/transactions';
 import { useMainStore } from '@/store/main';
 import { useNotifications } from '@/store/notifications';
 import { useReports } from '@/store/reports';
