+++ frontend/app/src/views/history/transactions/Transactions.vue
@@ -16,7 +16,7 @@ import { defineAsyncComponent, onBeforeMount, onUnmounted, watch } from 'vue';
 import { useI18n } from 'vue-i18n-composable';
 import { isSectionLoading, setupStatusChecking } from '@/composables/common';
 import { Section } from '@/store/const';
-import { useTransactions } from '@/store/history';
+import { useTransactions } from '@/store/history/transactions';
 import { useTasks } from '@/store/tasks';
 import { TaskType } from '@/types/task-type';
